requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

appdirs==1.4.4

beautifulsoup4==4.9.3

black==20.8b1

cachetools==4.2.1; python_version ~= '3.5'

certifi==2020.12.5

chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

et-xmlfile==1.0.1

google-auth-oauthlib==0.4.4; python_version >= '3.6'

google-auth==1.28.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

gspread==3.7.0

idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

lxml==4.6.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

multitasking==0.0.9

mypy-extensions==0.4.3

numpy==1.20.2; python_version >= '3.7'

oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

openpyxl==3.0.7

pandas==1.2.3; python_full_version >= '3.7.1'

pathspec==0.8.1

pillow==8.2.0

pyasn1-modules==0.2.8

pyasn1==0.4.8

python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

python-dotenv==0.17.0

pytz==2021.1

regex==2021.4.4

requests-oauthlib==1.3.0

requests==2.25.1

rsa==4.7.2; python_version >= '3.6'

six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

soupsieve==2.2.1; python_version >= '3.0'

toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'

typed-ast==1.4.2

typing-extensions==3.7.4.3

urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

yahoofinancials==1.6

yfinance==0.1.59

